requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
python-dotenv
requests
